requirements.txt
requests>=2.31.0
sseclient-py>=1.7.2
